Use HTTP Basic Auth when downloading MaxMind databases (#54)
* Adds HTTP Basic Auth for MaxMind downloads

It appears that MaxMind has changed the URLs for downloading databases, and
also wants authentication to happen via HTTP Basic Auth, rather than by
appending a "license_key" query string parameter. The old way of doing things
appears to still work, but we have no idea when it may start to fail.

This commit adds HTTP Basic Auth for MaxMind and updates the URLs for MaxMind
databases to the ones that MaxMind wants us to use.

* Updates maxmind_test.go to use HTTP Basic Auth

* Adds MAXMIND_ACCOUNT_ID to cloudbuild config and deploy script