Increases the avg download interval from 8h to 24h
The downloader runs in 3 projects and downloads 6 files (GeoLite Country, City
and ASN, in both binary and CSV format). With an average interval of 8h, this
means that each project may be downloading the database around 18 times per
day. Across 3 projects that is 54 downloads per day. The MaxMind download limit
for GeoLite accounts is only 30. This commit changes the average download
interval to 24h. The databases we use are only updated twice a week (Tue and
Fri):

https://support.maxmind.com/hc/en-us/articles/4408216129947-Download-and-Update-Databases

I do no know how this was working before my (recent
changes)[https://github.com/m-lab/downloader/pull/54]. I suspect that download
limit accounting may not have been in effect for the older URLs we were using,
and that when I updated the URLs this changed. Or possibly the older URLs
represented a MaxMind policy that had different limits. In any case, there are
strict download limits for the databases on the new URLs, and we are hitting it
every day, causing downloader to fail.