Removes test for MaxMind URLs
All the test did was to perform a HEAD request for the actual URL. However,
MaxMind logs the request and increments the daily limit counter for downloading
databases (which has a max of 30 per day). If something is wrong in the code
causing a build to fail and unit tests get run numerous times, testing alone
can exhaust the daily quota, even though nothing has actually been downloaded.
And this affects all projects. So some debugging or adding a feature in sandbox
could exhaust the daily quota, causing the production downloader to start
failing. This is counterproduction, and its not clear that we really benefit
from running a HEAD request of every MaxMind URL on every test run.